Update timestamp column logic * add column once when acc is created * move column to constants file
