--- extension/src/experiments/columns/collect/timestamp.ts
@@ -1,14 +1,0 @@
-import { ColumnAccumulator } from './util'
-import { ColumnType } from '../../webview/contract'
-
-export const timestampColumn = {
-  hasChildren: false,
-  label: 'Timestamp',
-  parentPath: ColumnType.TIMESTAMP,
-  path: 'Timestamp',
-  type: ColumnType.TIMESTAMP
-}
-
-export const collectTimestamp = (acc: ColumnAccumulator) => {
-  acc.timestamp = timestampColumn
-}
