+++ extension/src/experiments/columns/constants.ts
@@ -1,0 +1,9 @@
+import { ColumnType } from '../webview/contract'
+
+export const timestampColumn = {
+  hasChildren: false,
+  label: 'Timestamp',
+  parentPath: ColumnType.TIMESTAMP,
+  path: 'Timestamp',
+  type: ColumnType.TIMESTAMP
+}
